[midasexamples] Factor out some testutils from TutorialSuite
