+++ sim/firesim-lib/src/test/scala/TestSuiteCommon.scala
@@ -2,10 +2,21 @@
 package firesim
 
 import java.io.File
+import scala.io.Source
 import scala.sys.process.{stringSeqToProcess, ProcessLogger}
 
 /**
-  * NB: not thread-safe
+  * An base class for implementing FireSim integration tests that call out to the Make 
+  * buildsystem. These tests typically have three steps whose results are tracked by scalatest:
+  * 1) Elaborate the target and compile it through golden gate (ElaborateAndCompile)
+  * 2) Compile a metasimulator for the generated RTL (compileMlSimulator)
+  * 3) Run the metasimulator. Running a metasimualtion is somewaht
+  *    target-specific and is handled differently by different subclasses.
+  *
+  * Some tests inspect the simulation outputs, or run other simulators. See the
+  * [[TutorialSuite]] for examples of that.
+  *
+  * NB: Not thread-safe.
   */
 abstract class TestSuiteCommon extends org.scalatest.flatspec.AnyFlatSpec {
 
@@ -98,9 +109,48 @@ abstract class TestSuiteCommon extends org.scalatest.flatspec.AnyFlatSpec {
       }
     }
   }
+
+  /**
+    * Extracts all lines in a file that begin with a specific prefix, removing
+    * extra whitespace between the prefix and the remainder of the line
+    *
+    * @param filename Input file
+    * @param prefix The per-line prefix to filter with
+    * @param linesToDrop Some number of matched lines to be removed
+    * @param headerLines An initial number of lines to drop before filtering.
+    *        Assertions, Printf output have a single line header.
+    *        MLsim stdout has some unused output, so set this to 1 by default
+    *
+    */
+  def extractLines(filename: File, prefix: String, linesToDrop: Int = 0, headerLines: Int = 1): Seq[String] = {
+    val lines = Source.fromFile(filename).getLines.toList.drop(headerLines)
+    lines.filter(_.startsWith(prefix))
+         .dropRight(linesToDrop)
+         .map(_.stripPrefix(prefix).replaceAll(" +", " "))
+  }
+
+
+  /**
+    * Diffs two sets of lines. Wrap calls to this function in a scalatest
+    * behavior spec. @param aName and @param bName can be used to provide more
+    * insightful assertion messages in scalatest reporting.
+    */
+  def diffLines(
+      aLines: Seq[String],
+      bLines: Seq[String],
+      aName: String = "Actual output",
+      bName: String = "Expected output"): Unit = {
+    assert(aLines.size == bLines.size && aLines.nonEmpty,
+      s"\n${aName} length (${aLines.size}) and ${bName} length (${bLines.size}) differ.")
+    for ((a, b) <- bLines.zip(aLines)) {
+      assert(a == b)
+    }
+  }
 }
 
-// HACK: Hijacks TestSuiteCommon to run the MIDAS unit tests
+/**
+  * Hijacks TestSuiteCommon (mostly for make related features) to run the synthesizable unit tests.
+  */
 abstract class MidasUnitTestSuite(unitTestConfig: String, shouldFail: Boolean = false) extends TestSuiteCommon {
   val targetTuple = unitTestConfig
   // GENERATED_DIR & OUTPUT_DIR are only used to properly invoke `make clean`
